+++ founderz-common/src/main/java/com/founderz/common/vo/tag/TagId.java
@@ -1,7 +1,8 @@
-package com.founderz.common.vo;
+package com.founderz.common.vo.tag;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 import com.founderz.common.crypto.CryptoUtils;
+import com.founderz.common.vo.ValueObject;
 
 public record TagId(
         @JsonValue
