+++ founderz-common/src/main/java/com/founderz/internal/data/tag/TagDto.java
@@ -1,7 +1,7 @@
 package com.founderz.internal.data.tag;
 
-import com.founderz.common.vo.TagId;
-import com.founderz.common.vo.TagName;
+import com.founderz.common.vo.tag.TagId;
+import com.founderz.common.vo.tag.TagName;
 
 public record TagDto(
         TagId id,
